feat: add PublishThesisDto and update index export for thesis DTOs
