+++ src/domains/theses/dto/index.ts
@@ -1,3 +1,4 @@
 export * from '@/theses/dto/create-thesis.dto';
 export * from '@/theses/dto/review-thesis.dto';
 export * from '@/theses/dto/update-thesis.dto';
+export * from '@/theses/dto/publish-theses.dto';
