+++ src/domains/theses/dto/publish-theses.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsArray, IsBoolean, ValidateNested } from 'class-validator';
+
+export class PublishThesisDto {
+	@ApiProperty({ required: true })
+	@IsArray()
+	@ValidateNested({ each: true })
+	thesesIds: string[];
+
+	@ApiProperty({ required: true })
+	@IsBoolean()
+	isPublish: boolean;
+}
